+++ utils/src/main/kotlin/com/mirego/compose/utils/Configuration.kt
@@ -1,0 +1,12 @@
+package com.mirego.compose.utils
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun screenWidth(): Dp = LocalConfiguration.current.screenWidthDp.dp
+
+@Composable
+fun screenHeight(): Dp = LocalConfiguration.current.screenHeightDp.dp
