+++ utils/src/main/kotlin/com/mirego/compose/utils/extensions/DpExtensions.kt
@@ -1,0 +1,10 @@
+package com.mirego.compose.utils.extensions
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.unit.Dp
+
+@Composable
+fun Dp.toPx(): Float = with(LocalDensity.current) {
+    toPx()
+}
